v1.2
replaced ints by floats

+++ BytConverter.py
@@ -26,70 +26,70 @@ if s_unit in [b, B, kB, MB, GB, TB]:
             print("That's the same; "+ in_num + s_unit + " = " + in_num + d_unit + " ...")
         elif s_unit == b:
             if d_unit == B:
-                print("Result: ", (int(in_num) / 8), str(d_unit))
+                print("Result: ", (float(in_num) / 8), str(d_unit))
             elif d_unit == kB:
-                print("Result: ", ((int(in_num) / 8)/1024), str(d_unit))
+                print("Result: ", ((float(in_num) / 8)/1024), str(d_unit))
             elif d_unit == MB:
-                print("Result: ", (((int(in_num) / 8)/1024)/1024), str(d_unit))
+                print("Result: ", (((float(in_num) / 8)/1024)/1024), str(d_unit))
             elif d_unit == GB:
-                print("Result: ", ((((int(in_num) / 8) / 1024) / 1024) / 1024), str(d_unit))
+                print("Result: ", ((((float(in_num) / 8) / 1024) / 1024) / 1024), str(d_unit))
             elif d_unit == TB:
-                print("Result: ", (((((int(in_num) / 8) / 1024) / 1024) / 1024) / 1024), str(d_unit))
+                print("Result: ", (((((float(in_num) / 8) / 1024) / 1024) / 1024) / 1024), str(d_unit))
         elif s_unit == B:
             if d_unit == b:
-                print("Result: ", (int(in_num) * 8), str(d_unit))
+                print("Result: ", (float(in_num) * 8), str(d_unit))
             elif d_unit == kB:
-                print("Result: ", (int(in_num) / 1024), str(d_unit))
+                print("Result: ", (float(in_num) / 1024), str(d_unit))
             elif d_unit == MB:
-                print("Result: ", ((int(in_num) / 1024 / 1024), str(d_unit)))
+                print("Result: ", ((float(in_num) / 1024 / 1024), str(d_unit)))
             elif d_unit == GB:
-                print("Result: ", (((int(in_num) / 1024) / 1024) / 1024), str(d_unit))
+                print("Result: ", (((float(in_num) / 1024) / 1024) / 1024), str(d_unit))
             elif d_unit == TB:
-                print("Result: ", ((((int(in_num) / 1024) / 1024) / 1024) / 1024), str(d_unit))
+                print("Result: ", ((((float(in_num) / 1024) / 1024) / 1024) / 1024), str(d_unit))
         elif s_unit == kB:
             if d_unit == b:
-                print("Result: ", ((int(in_num) * 1024) * 8), str(d_unit))
+                print("Result: ", ((float(in_num) * 1024) * 8), str(d_unit))
             elif d_unit == B:
-                print("Result: ", (int(in_num) * 1024), str(d_unit))
+                print("Result: ", (float(in_num) * 1024), str(d_unit))
             elif d_unit == MB:
-                print("Result: ", ((int(in_num) / 1024), str(d_unit)))
+                print("Result: ", ((float(in_num) / 1024), str(d_unit)))
             elif d_unit == GB:
-                print("Result: ", ((int(in_num) / 1024) / 1024), str(d_unit))
+                print("Result: ", ((float(in_num) / 1024) / 1024), str(d_unit))
             elif d_unit == TB:
-                print("Result: ", (((int(in_num) / 1024) / 1024) / 1024), str(d_unit))
+                print("Result: ", (((float(in_num) / 1024) / 1024) / 1024), str(d_unit))
         elif s_unit == MB:
             if d_unit == b:
-                print("Result: ", (((int(in_num) * 1024) * 1024) * 8), str(d_unit))
+                print("Result: ", (((float(in_num) * 1024) * 1024) * 8), str(d_unit))
             elif d_unit == B:
-                print("Result: ", ((int(in_num) * 1024) * 1024), str(d_unit))
+                print("Result: ", ((float(in_num) * 1024) * 1024), str(d_unit))
             elif d_unit == kB:
-                print("Result: ", ((int(in_num) * 1024), str(d_unit)))
+                print("Result: ", ((float(in_num) * 1024), str(d_unit)))
             elif d_unit == GB:
-                print("Result: ", (int(in_num) / 1024), str(d_unit))
+                print("Result: ", (float(in_num) / 1024), str(d_unit))
             elif d_unit == TB:
-                print("Result: ", ((int(in_num) / 1024) / 1024), str(d_unit))
+                print("Result: ", ((float(in_num) / 1024) / 1024), str(d_unit))
         elif s_unit == GB:
             if d_unit == b:
-                print("Result: ", ((((int(in_num) * 1024) * 1024) * 1024) * 8), str(d_unit))
+                print("Result: ", ((((float(in_num) * 1024) * 1024) * 1024) * 8), str(d_unit))
             elif d_unit == B:
-                print("Result: ", (((int(in_num) * 1024) * 1024) * 1024), str(d_unit))
+                print("Result: ", (((float(in_num) * 1024) * 1024) * 1024), str(d_unit))
             elif d_unit == kB:
-                print("Result: ", (((int(in_num) * 1024) * 1024), str(d_unit)))
+                print("Result: ", ((float(in_num) * 1024) * 1024), str(d_unit))
             elif d_unit == MB:
-                print("Result: ", (int(in_num) * 1024), str(d_unit))
+                print("Result: ", (float(in_num) * 1024), str(d_unit))
             elif d_unit == TB:
-                print("Result: ", (int(in_num) / 1024), str(d_unit))
+                print("Result: ", (float(in_num) / 1024), str(d_unit))
         elif s_unit == TB:
             if d_unit == b:
-                print("Result: ", (((((int(in_num) * 1024) * 1024) * 1024) * 1024) * 8), str(d_unit))
+                print("Result: ", (((((float(in_num) * 1024) * 1024) * 1024) * 1024) * 8), str(d_unit))
             elif d_unit == B:
-                print("Result: ", ((((int(in_num) * 1024) * 1024) * 1024) * 1024), str(d_unit))
+                print("Result: ", ((((float(in_num) * 1024) * 1024) * 1024) * 1024), str(d_unit))
             elif d_unit == kB:
-                print("Result: ", ((((int(in_num) * 1024) * 1024) * 1024), str(d_unit)))
+                print("Result: ", ((((float(in_num) * 1024) * 1024) * 1024), str(d_unit)))
             elif d_unit == MB:
-                print("Result: ", ((int(in_num) * 1024) * 1024), str(d_unit))
+                print("Result: ", ((float(in_num) * 1024) * 1024), str(d_unit))
             elif d_unit == GB:
-                print("Result: ", (int(in_num) * 1024), str(d_unit))
+                print("Result: ", (float(in_num) * 1024), str(d_unit))
         print("")
         input("Press ENTER to exit")
     else: print("Bad Unit! Exiting...")
